+++ inst/include/dplyr/RowwiseDataFrame.h
@@ -1,7 +1,7 @@
 #ifndef dplyr_tools_RowwiseDataFrame_H
 #define dplyr_tools_RowwiseDataFrame_H
 
-namespace Rcpp {
+namespace dplyr {
 
   class RowwiseDataFrame;
 
@@ -75,6 +75,11 @@ namespace Rcpp {
     IntegerVector group_sizes;
 
   };
+
+}
+
+namespace Rcpp {
+  using namespace dplyr;
 
   template <>
   inline bool is<RowwiseDataFrame>(SEXP x) {
